bugs

+++ sish.c
@@ -6,7 +6,7 @@
 
 
 
-int MAX = 1024; // max length
+size_t MAX = 1024; // max length
 int flag = 1;
 
 void run(char *args[])
